+++ src/auth/pages/sign-up-or-login-page/components/sign-up-form/sign-up-form.component.html
@@ -87,7 +87,7 @@
     </div>
     <div>
       <button
-        (click)="onEmailAndPasswordSignUp()"
+        (click)="onSubmit()"
         type="submit"
         class="w-full bg-black text-white p-2 rounded-md hover:bg-gray-800 focus:outline-none focus:bg-black focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-gray-900 transition-colors duration-300">
         Sign Up
